requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
Pillow>=9.0.0
matplotlib>=3.5.0
